Fixed error in computing block numbers (was int overflow problem).
Added a super instruction check_block_number_calculation and test
that exercises it src/sialx/test/check_block_number_calc.sials to test_sial.

There are two "systems" in the test.  A small one, that is not very useful,
and a large one that reproduced the system where the bug showed
up at scale.  The latter is too big to run as a matter of course and is
commented out.

+++ src/sip/static_data/array_table.cpp
@@ -73,7 +73,7 @@ void ArrayTableEntry::init_calculated_values(const IndexTable& index_table){
 }
 
 size_t ArrayTableEntry::block_number(const BlockId& id) const{
-	int res = 0;
+	size_t res = 0;
 	for (int i = 0; i < rank_; i++){
 		res += (slice_sizes_[i] * (id.index_values(i) - lower_[i]));
 	}
